Struct rusoto_route53domains::DomainSuggestion

pub struct DomainSuggestion {
    pub availability: Option<String>,
    pub domain_name: Option<String>,
}

Information about one suggested domain name.

Fields
availability: Option<String>

Whether the domain name is available for registering.

You can register only the domains that are designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLERESERVED

The domain name is reserved under specific conditions.

AVAILABLEPREORDER

The domain name is available and can be preordered.

DONTKNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Amazon Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLEPREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

domain_name: Option<String>

A suggested domain name.