Struct rusoto_route53domains::RegisterDomainRequest

pub struct RegisterDomainRequest {
    pub admin_contact: ContactDetail,
    pub auto_renew: Option<bool>,
    pub domain_name: String,
    pub duration_in_years: i64,
    pub idn_lang_code: Option<String>,
    pub privacy_protect_admin_contact: Option<bool>,
    pub privacy_protect_registrant_contact: Option<bool>,
    pub privacy_protect_tech_contact: Option<bool>,
    pub registrant_contact: ContactDetail,
    pub tech_contact: ContactDetail,
}

The RegisterDomain request includes the following elements.

Fields
admin_contact: ContactDetail

Provides detailed contact information.

auto_renew: Option<bool>

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

domain_name: String

The domain name that you want to register.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

duration_in_years: i64

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

idn_lang_code: Option<String>

Reserved for future use.

privacy_protect_admin_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

privacy_protect_registrant_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

privacy_protect_tech_contact: Option<bool>

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

registrant_contact: ContactDetail

Provides detailed contact information.

tech_contact: ContactDetail

Provides detailed contact information.