Trait rusoto_route53domains::Route53Domains

pub trait Route53Domains {
    fn check_domain_availability(
        &self, 
        input: CheckDomainAvailabilityRequest
    ) -> RusotoFuture<CheckDomainAvailabilityResponse, CheckDomainAvailabilityError>;

    fn check_domain_transferability(
        &self, 
        input: CheckDomainTransferabilityRequest
    ) -> RusotoFuture<CheckDomainTransferabilityResponse, CheckDomainTransferabilityError>;

    fn delete_tags_for_domain(
        &self, 
        input: DeleteTagsForDomainRequest
    ) -> RusotoFuture<DeleteTagsForDomainResponse, DeleteTagsForDomainError>;

    fn disable_domain_auto_renew(
        &self, 
        input: DisableDomainAutoRenewRequest
    ) -> RusotoFuture<DisableDomainAutoRenewResponse, DisableDomainAutoRenewError>;

    fn disable_domain_transfer_lock(
        &self, 
        input: DisableDomainTransferLockRequest
    ) -> RusotoFuture<DisableDomainTransferLockResponse, DisableDomainTransferLockError>;

    fn enable_domain_auto_renew(
        &self, 
        input: EnableDomainAutoRenewRequest
    ) -> RusotoFuture<EnableDomainAutoRenewResponse, EnableDomainAutoRenewError>;

    fn enable_domain_transfer_lock(
        &self, 
        input: EnableDomainTransferLockRequest
    ) -> RusotoFuture<EnableDomainTransferLockResponse, EnableDomainTransferLockError>;

    fn get_contact_reachability_status(
        &self, 
        input: GetContactReachabilityStatusRequest
    ) -> RusotoFuture<GetContactReachabilityStatusResponse, GetContactReachabilityStatusError>;

    fn get_domain_detail(
        &self, 
        input: GetDomainDetailRequest
    ) -> RusotoFuture<GetDomainDetailResponse, GetDomainDetailError>;

    fn get_domain_suggestions(
        &self, 
        input: GetDomainSuggestionsRequest
    ) -> RusotoFuture<GetDomainSuggestionsResponse, GetDomainSuggestionsError>;

    fn get_operation_detail(
        &self, 
        input: GetOperationDetailRequest
    ) -> RusotoFuture<GetOperationDetailResponse, GetOperationDetailError>;

    fn list_domains(
        &self, 
        input: ListDomainsRequest
    ) -> RusotoFuture<ListDomainsResponse, ListDomainsError>;

    fn list_operations(
        &self, 
        input: ListOperationsRequest
    ) -> RusotoFuture<ListOperationsResponse, ListOperationsError>;

    fn list_tags_for_domain(
        &self, 
        input: ListTagsForDomainRequest
    ) -> RusotoFuture<ListTagsForDomainResponse, ListTagsForDomainError>;

    fn register_domain(
        &self, 
        input: RegisterDomainRequest
    ) -> RusotoFuture<RegisterDomainResponse, RegisterDomainError>;

    fn renew_domain(
        &self, 
        input: RenewDomainRequest
    ) -> RusotoFuture<RenewDomainResponse, RenewDomainError>;

    fn resend_contact_reachability_email(
        &self, 
        input: ResendContactReachabilityEmailRequest
    ) -> RusotoFuture<ResendContactReachabilityEmailResponse, ResendContactReachabilityEmailError>;

    fn retrieve_domain_auth_code(
        &self, 
        input: RetrieveDomainAuthCodeRequest
    ) -> RusotoFuture<RetrieveDomainAuthCodeResponse, RetrieveDomainAuthCodeError>;

    fn transfer_domain(
        &self, 
        input: TransferDomainRequest
    ) -> RusotoFuture<TransferDomainResponse, TransferDomainError>;

    fn update_domain_contact(
        &self, 
        input: UpdateDomainContactRequest
    ) -> RusotoFuture<UpdateDomainContactResponse, UpdateDomainContactError>;

    fn update_domain_contact_privacy(
        &self, 
        input: UpdateDomainContactPrivacyRequest
    ) -> RusotoFuture<UpdateDomainContactPrivacyResponse, UpdateDomainContactPrivacyError>;

    fn update_domain_nameservers(
        &self, 
        input: UpdateDomainNameserversRequest
    ) -> RusotoFuture<UpdateDomainNameserversResponse, UpdateDomainNameserversError>;

    fn update_tags_for_domain(
        &self, 
        input: UpdateTagsForDomainRequest
    ) -> RusotoFuture<UpdateTagsForDomainResponse, UpdateTagsForDomainError>;

    fn view_billing(
        &self, 
        input: ViewBillingRequest
    ) -> RusotoFuture<ViewBillingResponse, ViewBillingError>;
}

Trait representing the capabilities of the Amazon Route 53 Domains API. Amazon Route 53 Domains clients implement this trait.

Required methods

fn check_domain_availability(
&self,
input: CheckDomainAvailabilityRequest
) -> RusotoFuture<CheckDomainAvailabilityResponse, CheckDomainAvailabilityError>

This operation checks the availability of one domain name. Note that if the availability status of a domain is pending, you must submit another request to determine the availability of the domain name.

fn check_domain_transferability(
&self,
input: CheckDomainTransferabilityRequest
) -> RusotoFuture<CheckDomainTransferabilityResponse, CheckDomainTransferabilityError>

Checks whether a domain name can be transferred to Amazon Route 53.

fn delete_tags_for_domain(
&self,
input: DeleteTagsForDomainRequest
) -> RusotoFuture<DeleteTagsForDomainResponse, DeleteTagsForDomainError>

This operation deletes the specified tags for a domain.

All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.

fn disable_domain_auto_renew(
&self,
input: DisableDomainAutoRenewRequest
) -> RusotoFuture<DisableDomainAutoRenewResponse, DisableDomainAutoRenewError>

This operation disables automatic renewal of domain registration for the specified domain.

fn disable_domain_transfer_lock(
&self,
input: DisableDomainTransferLockRequest
) -> RusotoFuture<DisableDomainTransferLockResponse, DisableDomainTransferLockError>

This operation removes the transfer lock on the domain (specifically the clientTransferProhibited status) to allow domain transfers. We recommend you refrain from performing this action unless you intend to transfer the domain to a different registrar. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn enable_domain_auto_renew(
&self,
input: EnableDomainAutoRenewRequest
) -> RusotoFuture<EnableDomainAutoRenewResponse, EnableDomainAutoRenewError>

This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires. The cost of renewing your domain registration is billed to your AWS account.

The period during which you can renew a domain name varies by TLD. For a list of TLDs and their renewal policies, see "Renewal, restoration, and deletion times" on the website for our registrar associate, Gandi. Amazon Route 53 requires that you renew before the end of the renewal period that is listed on the Gandi website so we can complete processing before the deadline.

fn enable_domain_transfer_lock(
&self,
input: EnableDomainTransferLockRequest
) -> RusotoFuture<EnableDomainTransferLockResponse, EnableDomainTransferLockError>

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn get_contact_reachability_status(
&self,
input: GetContactReachabilityStatusRequest
) -> RusotoFuture<GetContactReachabilityStatusResponse, GetContactReachabilityStatusError>

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded.

If you want us to resend the email, use the ResendContactReachabilityEmail operation.

fn get_domain_detail(
&self,
input: GetDomainDetailRequest
) -> RusotoFuture<GetDomainDetailResponse, GetDomainDetailError>

This operation returns detailed information about a specified domain that is associated with the current AWS account. Contact information for the domain is also returned as part of the output.

fn get_domain_suggestions(
&self,
input: GetDomainSuggestionsRequest
) -> RusotoFuture<GetDomainSuggestionsResponse, GetDomainSuggestionsError>

The GetDomainSuggestions operation returns a list of suggested domain names given a string, which can either be a domain name or simply a word or phrase (without spaces).

fn get_operation_detail(
&self,
input: GetOperationDetailRequest
) -> RusotoFuture<GetOperationDetailResponse, GetOperationDetailError>

This operation returns the current status of an operation that is not completed.

fn list_domains(
&self,
input: ListDomainsRequest
) -> RusotoFuture<ListDomainsResponse, ListDomainsError>

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

fn list_operations(
&self,
input: ListOperationsRequest
) -> RusotoFuture<ListOperationsResponse, ListOperationsError>

This operation returns the operation IDs of operations that are not yet complete.

fn list_tags_for_domain(
&self,
input: ListTagsForDomainRequest
) -> RusotoFuture<ListTagsForDomainResponse, ListTagsForDomainError>

This operation returns all of the tags that are associated with the specified domain.

All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.

fn register_domain(
&self,
input: RegisterDomainRequest
) -> RusotoFuture<RegisterDomainResponse, RegisterDomainError>

This operation registers a domain. Domains are registered either by Amazon Registrar (for .com, .net, and .org domains) or by our registrar associate, Gandi (for all other domains). For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.

Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.

Optionally enables privacy protection, so WHOIS queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you don't enable privacy protection, WHOIS queries return the information that you entered for the registrant, admin, and tech contacts.

If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.

Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

fn renew_domain(
&self,
input: RenewDomainRequest
) -> RusotoFuture<RenewDomainResponse, RenewDomainError>

This operation renews a domain for the specified number of years. The cost of renewing your domain is billed to your AWS account.

We recommend that you renew your domain several weeks before the expiration date. Some TLD registries delete domains before the expiration date if you haven't renewed far enough in advance. For more information about renewing domain registration, see Renewing Registration for a Domain in the Amazon Route 53 Developer Guide.

fn resend_contact_reachability_email(
&self,
input: ResendContactReachabilityEmailRequest
) -> RusotoFuture<ResendContactReachabilityEmailResponse, ResendContactReachabilityEmailError>

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation resends the confirmation email to the current email address for the registrant contact.

fn retrieve_domain_auth_code(
&self,
input: RetrieveDomainAuthCodeRequest
) -> RusotoFuture<RetrieveDomainAuthCodeResponse, RetrieveDomainAuthCodeError>

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

fn transfer_domain(
&self,
input: TransferDomainRequest
) -> RusotoFuture<TransferDomainResponse, TransferDomainError>

This operation transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered either with Amazon Registrar (for .com, .net, and .org domains) or with our registrar associate, Gandi (for all other TLDs).

For transfer requirements, a detailed procedure, and information about viewing the status of a domain transfer, see Transferring Registration for a Domain to Amazon Route 53 in the Amazon Route 53 Developer Guide.

If the registrar for your domain is also the DNS service provider for the domain, we highly recommend that you consider transferring your DNS service to Amazon Route 53 or to another DNS service provider before you transfer your registration. Some registrars provide free DNS service when you purchase a domain registration. When you transfer the registration, the previous registrar will not renew your domain registration and could end your DNS service at any time.

If the registrar for your domain is also the DNS service provider for the domain and you don't transfer DNS service to another provider, your website, email, and the web applications associated with the domain might become unavailable.

If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the transfer doesn't complete successfully, the domain registrant will be notified by email.

fn update_domain_contact(
&self,
input: UpdateDomainContactRequest
) -> RusotoFuture<UpdateDomainContactResponse, UpdateDomainContactError>

This operation updates the contact information for a particular domain. You must specify information for at least one contact: registrant, administrator, or technical.

If the update is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn update_domain_contact_privacy(
&self,
input: UpdateDomainContactPrivacyRequest
) -> RusotoFuture<UpdateDomainContactPrivacyResponse, UpdateDomainContactPrivacyError>

This operation updates the specified domain contact's privacy setting. When privacy protection is enabled, contact information such as email address is replaced either with contact information for Amazon Registrar (for .com, .net, and .org domains) or with contact information for our registrar associate, Gandi.

This operation affects only the contact information for the specified contact type (registrant, administrator, or tech). If the request succeeds, Amazon Route 53 returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request doesn't complete successfully, the domain registrant will be notified by email.

fn update_domain_nameservers(
&self,
input: UpdateDomainNameserversRequest
) -> RusotoFuture<UpdateDomainNameserversResponse, UpdateDomainNameserversError>

This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

fn update_tags_for_domain(
&self,
input: UpdateTagsForDomainRequest
) -> RusotoFuture<UpdateTagsForDomainResponse, UpdateTagsForDomainError>

This operation adds or updates tags for a specified domain.

All tag operations are eventually consistent; subsequent operations might not immediately represent all issued operations.

fn view_billing(
&self,
input: ViewBillingRequest
) -> RusotoFuture<ViewBillingResponse, ViewBillingError>

Returns all the domain-related billing records for the current AWS account for a specified period

Loading content...