Crate rusoto_route53domains

Amazon Route 53 API actions let you register domain names and perform related operations.

If you're using the service, you're probably looking for Route53DomainsClient and Route53Domains.

Structs

BillingRecord
Information for one billing record.

CheckDomainAvailabilityRequest
The CheckDomainAvailability request contains the following elements.

CheckDomainAvailabilityResponse
The CheckDomainAvailability response includes the following elements.

CheckDomainTransferabilityRequest
The CheckDomainTransferability request contains the following elements.

CheckDomainTransferabilityResponse
The CheckDomainTransferability response includes the following elements.

ContactDetail
ContactDetail includes the following elements.

DeleteTagsForDomainRequest
The DeleteTagsForDomainRequest includes the following elements.

DeleteTagsForDomainResponse

DisableDomainAutoRenewRequest

DisableDomainAutoRenewResponse

DisableDomainTransferLockRequest
The DisableDomainTransferLock request includes the following element.

DisableDomainTransferLockResponse
The DisableDomainTransferLock response includes the following element.

DomainSuggestion
Information about one suggested domain name.

DomainSummary
Summary information about one domain.

DomainTransferability
A complex type that contains information about whether the specified domain can be transferred to Amazon Route 53.

EnableDomainAutoRenewRequest

EnableDomainAutoRenewResponse

EnableDomainTransferLockRequest
A request to set the transfer lock for the specified domain.

EnableDomainTransferLockResponse
The EnableDomainTransferLock response includes the following elements.

ExtraParam
ExtraParam includes the following elements.

GetContactReachabilityStatusRequest

GetContactReachabilityStatusResponse

GetDomainDetailRequest
The GetDomainDetail request includes the following element.

GetDomainDetailResponse
The GetDomainDetail response includes the following elements.

GetDomainSuggestionsRequest

GetDomainSuggestionsResponse

GetOperationDetailRequest
The GetOperationDetail request includes the following element.

GetOperationDetailResponse
The GetOperationDetail response includes the following elements.

ListDomainsRequest
The ListDomains request includes the following elements.

ListDomainsResponse
The ListDomains response includes the following elements.

ListOperationsRequest
The ListOperations request includes the following elements.

ListOperationsResponse
The ListOperations response includes the following elements.

ListTagsForDomainRequest
The ListTagsForDomainRequest includes the following elements.

ListTagsForDomainResponse
The ListTagsForDomain response includes the following elements.

Nameserver
Nameserver includes the following elements.

OperationSummary
OperationSummary includes the following elements.

RegisterDomainRequest
The RegisterDomain request includes the following elements.

RegisterDomainResponse
The RegisterDomain response includes the following element.

RenewDomainRequest
A RenewDomain request includes the number of years that you want to renew for and the current expiration year.

RenewDomainResponse

ResendContactReachabilityEmailRequest

ResendContactReachabilityEmailResponse

RetrieveDomainAuthCodeRequest
A request for the authorization code for the specified domain. To transfer a domain to another registrar, you provide this value to the new registrar.

RetrieveDomainAuthCodeResponse
The RetrieveDomainAuthCode response includes the following element.

Route53DomainsClient
A client for the Amazon Route 53 Domains API.

Tag
Each tag includes the following elements.

TransferDomainRequest
The TransferDomain request includes the following elements.

TransferDomainResponse
The TranserDomain response includes the following element.

UpdateDomainContactPrivacyRequest
The UpdateDomainContactPrivacy request includes the following elements.

UpdateDomainContactPrivacyResponse
The UpdateDomainContactPrivacy response includes the following element.

UpdateDomainContactRequest
The UpdateDomainContact request includes the following elements.

UpdateDomainContactResponse
The UpdateDomainContact response includes the following element.

UpdateDomainNameserversRequest
Replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

UpdateDomainNameserversResponse
The UpdateDomainNameservers response includes the following element.

UpdateTagsForDomainRequest
The UpdateTagsForDomainRequest includes the following elements.

UpdateTagsForDomainResponse

ViewBillingRequest
The ViewBilling request includes the following elements.

ViewBillingResponse
The ViewBilling response includes the following elements.

Enums

CheckDomainAvailabilityError
Errors returned by CheckDomainAvailability

CheckDomainTransferabilityError
Errors returned by CheckDomainTransferability

DeleteTagsForDomainError
Errors returned by DeleteTagsForDomain

DisableDomainAutoRenewError
Errors returned by DisableDomainAutoRenew

DisableDomainTransferLockError
Errors returned by DisableDomainTransferLock

EnableDomainAutoRenewError
Errors returned by EnableDomainAutoRenew

EnableDomainTransferLockError
Errors returned by EnableDomainTransferLock

GetContactReachabilityStatusError
Errors returned by GetContactReachabilityStatus

GetDomainDetailError
Errors returned by GetDomainDetail

GetDomainSuggestionsError
Errors returned by GetDomainSuggestions

GetOperationDetailError
Errors returned by GetOperationDetail

ListDomainsError
Errors returned by ListDomains

ListOperationsError
Errors returned by ListOperations

ListTagsForDomainError
Errors returned by ListTagsForDomain

RegisterDomainError
Errors returned by RegisterDomain

RenewDomainError
Errors returned by RenewDomain

ResendContactReachabilityEmailError
Errors returned by ResendContactReachabilityEmail

RetrieveDomainAuthCodeError
Errors returned by RetrieveDomainAuthCode

TransferDomainError
Errors returned by TransferDomain

UpdateDomainContactError
Errors returned by UpdateDomainContact

UpdateDomainContactPrivacyError
Errors returned by UpdateDomainContactPrivacy

UpdateDomainNameserversError
Errors returned by UpdateDomainNameservers

UpdateTagsForDomainError
Errors returned by UpdateTagsForDomain

ViewBillingError
Errors returned by ViewBilling

Traits

Route53Domains
Trait representing the capabilities of the Amazon Route 53 Domains API. Amazon Route 53 Domains clients implement this trait.