Struct rusoto_route53domains::ContactDetail

pub struct ContactDetail {
    pub address_line_1: Option<String>,
    pub address_line_2: Option<String>,
    pub city: Option<String>,
    pub contact_type: Option<String>,
    pub country_code: Option<String>,
    pub email: Option<String>,
    pub extra_params: Option<Vec<ExtraParam>>,
    pub fax: Option<String>,
    pub first_name: Option<String>,
    pub last_name: Option<String>,
    pub organization_name: Option<String>,
    pub phone_number: Option<String>,
    pub state: Option<String>,
    pub zip_code: Option<String>,
}

ContactDetail includes the following elements.

Fields
address_line_1: Option<String>

First line of the contact's address.

address_line_2: Option<String>

Second line of contact's address, if any.

city: Option<String>

The city of the contact's address.

contact_type: Option<String>

Indicates whether the contact is a person, company, association, or public organization. If you choose an option other than PERSON, you must enter an organization name, and you can't enable privacy protection for the contact.

country_code: Option<String>

Code for the country of the contact's address.

email: Option<String>

Email address of the contact.

extra_params: Option<Vec<ExtraParam>>

A list of name-value pairs for parameters required by certain top-level domains.

fax: Option<String>

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

first_name: Option<String>

First name of contact.

last_name: Option<String>

Last name of contact.

organization_name: Option<String>

Name of the organization for contact types other than PERSON.

phone_number: Option<String>

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

state: Option<String>

The state or province of the contact's city.

zip_code: Option<String>

The zip or postal code of the contact's address.