Struct rusoto_route53domains::GetDomainDetailResponse

pub struct GetDomainDetailResponse {
    pub abuse_contact_email: Option<String>,
    pub abuse_contact_phone: Option<String>,
    pub admin_contact: ContactDetail,
    pub admin_privacy: Option<bool>,
    pub auto_renew: Option<bool>,
    pub creation_date: Option<f64>,
    pub dns_sec: Option<String>,
    pub domain_name: String,
    pub expiration_date: Option<f64>,
    pub nameservers: Vec<Nameserver>,
    pub registrant_contact: ContactDetail,
    pub registrant_privacy: Option<bool>,
    pub registrar_name: Option<String>,
    pub registrar_url: Option<String>,
    pub registry_domain_id: Option<String>,
    pub reseller: Option<String>,
    pub status_list: Option<Vec<String>>,
    pub tech_contact: ContactDetail,
    pub tech_privacy: Option<bool>,
    pub updated_date: Option<f64>,
    pub who_is_server: Option<String>,
}

The GetDomainDetail response includes the following elements.

Fields
abuse_contact_email: Option<String>

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

abuse_contact_phone: Option<String>

Phone number for reporting abuse.

admin_contact: ContactDetail

Provides details about the domain administrative contact.

admin_privacy: Option<bool>

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the admin contact.

auto_renew: Option<bool>

Specifies whether the domain registration is set to renew automatically.

creation_date: Option<f64>

The date when the domain was created as found in the response to a WHOIS query. The date and time is in Coordinated Universal time (UTC).

dns_sec: Option<String>

Reserved for future use.

domain_name: String

The name of a domain.

expiration_date: Option<f64>

The date when the registration for the domain is set to expire. The date and time is in Coordinated Universal time (UTC).

nameservers: Vec<Nameserver>

The name of the domain.

registrant_contact: ContactDetail

Provides details about the domain registrant.

registrant_privacy: Option<bool>

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

registrar_name: Option<String>

Name of the registrar of the domain as identified in the registry. Domains with a .com, .net, or .org TLD are registered by Amazon Registrar. All other domains are registered by our registrar associate, Gandi. The value for domains that are registered by Gandi is "GANDI SAS".

registrar_url: Option<String>

Web address of the registrar.

registry_domain_id: Option<String>

Reserved for future use.

reseller: Option<String>

Reseller of the domain. Domains registered or transferred using Amazon Route 53 domains will have "Amazon" as the reseller.

status_list: Option<Vec<String>>

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

tech_contact: ContactDetail

Provides details about the domain technical contact.

tech_privacy: Option<bool>

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the technical contact.

updated_date: Option<f64>

The last updated date of the domain as found in the response to a WHOIS query. The date and time is in Coordinated Universal time (UTC).

who_is_server: Option<String>

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.