Struct rusoto_route53domains::Route53DomainsClient

pub struct Route53DomainsClient { /* fields omitted */ }

A client for the Amazon Route 53 Domains API.

Methods

impl Route53DomainsClient

pub fn new(region: Region) -> Route53DomainsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Route53DomainsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,